Only operate on 'with' columns in commands

+++ src/services/commands/ensure-leading-command.ts
@@ -31,14 +31,26 @@ export class EnsureLeadingCommand implements Command {
 
             for (let i = 0; i < value.length; i++) {
 
-                result.push(this.services.textUtilsService.EnsureLeading(value[i], para));
+                if (context.withIndices.includes(i)) {
+
+                    result.push(this.services.textUtilsService.EnsureLeading(value[i], para));
+                }
+                else {
+                    result.push(value[i]);
+                }
             }
         }
         else {
 
             for (let i = 0; i < value.length; i++) {
 
-                result.push(this.services.textUtilsService.RemoveLeading(value[i], para));
+                if (context.withIndices.includes(i)) {
+                    
+                    result.push(this.services.textUtilsService.RemoveLeading(value[i], para));
+                }
+                else {
+                    result.push(value[i]);
+                }
             }
         }
